requirements.txt
flask
authlib>=0.10
requests
psycopg2-binary
pytz
gunicorn>=19.9.0
